requirements_temp.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
gitdb==4.0.8
GitPython==3.1.24
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyngrok==5.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.10.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
